Checkpoint 3
Most tests working: Fix some tests in: CatscriptParserStatementsTest and eval

+++ src/main/java/edu/montana/csci/csci468/parser/expressions/IntegerLiteralExpression.java
@@ -35,7 +35,7 @@ public class IntegerLiteralExpression extends Expression {
 
     @Override
     public Object evaluate(CatscriptRuntime runtime) {
-        return integerVal;
+        return (Integer) integerVal;
     }
 
     @Override
